# openapi.yaml
openapi: 3.1.0
info:
  title: Veilence-MX API
  description: >
    Enterprise SaaS supply chain compromise monitor for PyPI and npm packages.
    Provides endpoints for package monitoring, release analysis, alert management,
    organization/RBAC, notifications, and system administration.
  version: 1.0.0
  contact:
    name: Veilence-MX Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.veilence.example.com
    description: Production server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/orgs/{orgId}/audit-logs:
    get:
      tags: [Organizations]
      summary: List audit logs
      description: >
        Returns a paginated list of audit logs for the organization.
        Requires `audit:read` permission.
      operationId: listAuditLogs
      parameters:
        - $ref: "#/components/parameters/OrgIdPath"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: action
          in: query
          schema:
            type: string
        - name: resource
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: integer
        - name: from_date
          in: query
          schema:
            type: string
            format: date-time
        - name: to_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Paginated audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_AuditLogList"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── Notification Channels (org-scoped) ─────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login or /api/auth/register
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key created via /api/auth/api-keys
  parameters:
...
    OrgIdPath:
      name: orgId
      in: path
      required: true
      schema:
        type: integer
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        data:
          nullable: true
        error:
          type: string

    # ── Domain Schemas ──────────────────────────────────────────────
...
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        orgId:
          type: integer
        action:
          type: string
        resource:
          type: string
        resourceId:
          type: integer
        details:
          type: string
        ipAddress:
          type: string
        userAgent:
          type: string
        correlationId:
          type: string
        createdAt:
          type: string
          format: date-time
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
          format: int64

    # ── Request Schemas ─────────────────────────────────────────────
...
    ApiResponse_AuditLogList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        error:
          type: string
          nullable: true
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...